api callbacks added!
