+++ src/routes/api.js
@@ -1,26 +1,23 @@
-const express = require('express');
-const {executeQuery,insertDataIntoTable}= require('../utility')
-const connectToDatabase = require('../connection');
-
-const {
-  orderSummary,
-  topRestaurant,
-  getBusiestDelivery_Boy,
-  getfrequentOrderDetails,
-  getSpecificOrderPrice,
-  getSpecificOrdersDetails,
-  getOutOfDeliveryOrder,
-  getOrderAssignedDeliverBoy,
-  getTotalSalesForRestaurant,
-  getDeliveryBoyAssignedToSpecificOrder,
-  getDeliveryBoyStatus,
-  getActiveDeliveryBoys,
-  getDeliveryBoyDetailsForSpecificOrder,
-} = require('./sqlQuery');
-
+const express = require("express");
 const router = express.Router();
+const {
+  insertData,
+  updateData,
+  order_Summary,
+  getTopRestaurt,
+  getBusiestDeliveryBoy,
+  getFrequestOrder,
+  getSpecificOrderById,
+  getDeliveryBoyByOrderId,
+  getOrderDetailsByOrderId,
+  getOutForDeliveryDetails,
+  assignedDeliveryBoyByorderId,
+  getTotalSalesByRestaurantId,
+  getDeliveryBoy,
+  getActiveDeliveryBoy,
+  getdeliveryboyDetails,
+} = require("./apiCallbacks");
 
-const validTables = ['restaurants', 'customers', 'delivery_boys', 'orders', 'order_items']; // Define valid table names
 /**
  * @swagger
  * /{entityName}:
@@ -53,55 +50,12 @@ const validTables = ['restaurants', 'customers', 'delivery_boys', 'orders', 'ord
  *       500:
  *         description: Internal Server Error
  */
-router.post('/:entityName', async (req, res) => {
-  const { entityName } = req.params;
-  let data = req.body;
-
-
-  console.log('Incoming data:', data);
-
-  // Validating table name
-  if (!validTables.includes(entityName)) {
-    return res.status(400).json({ error: `Invalid table name: ${entityName}` });
-  }
-
-  
-  if (!Array.isArray(data)) {
-    data = [data]; 
-  }
-
-  // Validate data array
-  if (data.length === 0 || !data.every((item) => typeof item === 'object')) {
-    return res.status(400).json({ error: 'Invalid data format. Expected JSON objects ' });
-  }
-
-  const columns = Object.keys(data[0]);
-  console.log('Columns:', columns);
-
-  if (columns.length === 0) {
-    return res.status(400).json({ error: 'No data provided to insert.' });
-  }
-
-  try {
-    const connection = await connectToDatabase();
-
-    // Insert each item into the table
-    await insertDataIntoTable(connection, entityName, data, columns);
-
-    res.status(201).json({ message: `Data successfully inserted into ${entityName}` });
-
-    connection.end(); // Close the connection
-  } catch (error) {
-    console.error('Error inserting data:', error);
-    res.status(500).json({ error: 'Failed to insert data', details: error.message });
-  }
-});
-
+router.post("/:entityName", insertData);
 /**
  * @swagger
  * /{entityName}/{id}:
  *   put:
- *     summary: Update data into tables (restaurants, customers, orders, delivery_boys, order_items)
+ *     summary: Update data for tables (restaurants, customers, orders, delivery_boys, order_items)
  *     description: Update data for the specified entity (table) by ID
  *     parameters:
  *       - name: entityName
@@ -131,68 +85,7 @@ router.post('/:entityName', async (req, res) => {
  *       500:
  *         description: Internal Server Error
  */
-router.put('/:entityName/:id', async (req, res) => {
-  const { entityName, id } = req.params;
-  let data = req.body;
-
-  
-  console.log('Incoming data:', data);
-
-  
-  if (!validTables.includes(entityName)) {
-    return res.status(400).json({ error: `Invalid table name: ${entityName}` });
-  }
-
-  
-  if (!Array.isArray(data)) {
-    return res.status(400).json({ error: 'Data must be an array of objects for updating.' });
-  }
-
-  // If the array has only one object, process that object
-  if (data.length === 1) {
-    data = data[0]; // Extract the first object
-  }
-
-  // Validate data object
-  if (typeof data !== 'object' || Object.keys(data).length === 0) {
-    return res.status(400).json({
-      error: 'Invalid data format. Expected a JSON object with key-value pairs for updating.',
-    });
-  }
-
-  // Prepare columns and values for the SQL query
-  const updates = Object.keys(data)
-    .map((key) => `\`${key}\` = ?`) // skip column names with backticks
-    .join(', ');  // Join all the updates with commas
-  const values = Object.values(data);
-
-  values.push(id);
-
-  try {
-    const connection = await connectToDatabase();
-
-    const query = `UPDATE \`${entityName}\` SET ${updates} WHERE id = ?`;
-
-    console.log('Generated Query:', query);
-    console.log('Values:', values);
-
-    const [result] = await connection.execute(query, values);
-
-    // Close the connection
-    connection.end();
-
-    // Respond based on the result
-    if (result.affectedRows === 0) {
-      return res.status(404).json({ error: `No record found with id: ${id} in ${entityName}` });
-    }
-
-    res.status(200).json({ message: `Record with id ${id} successfully updated in ${entityName}` });
-  } catch (error) {
-    console.error('Error updating data:', error);
-    res.status(500).json({ error: 'Failed to update data', details: error.message });
-  }
-});
-
+router.put("/:entityName/:id", updateData);
 /**
  * @swagger
  * /order-summary:
@@ -220,23 +113,9 @@ router.put('/:entityName/:id', async (req, res) => {
  *       500:
  *         description: Failed to fetch order summary
  */
-router.get('/order-summary', async (req, res) => {
-    let { limit , offset} = req.query;
+router.get("/order-summary", order_Summary);
 
-    limit = (isNaN(limit) || limit <= 0) ? '10' : String(limit); // setting default to 10 
-    offset = (isNaN(offset) || offset < 0) ? '0' : String(offset); // setting default to 0
-  
-    try {
-      
-      const result = await executeQuery(orderSummary, [limit, offset]);
-      res.status(200).json(result);  
-    } catch (error) {
-      console.error('Failed to fetch order summary:', error.message);
-      res.status(500).send({ error: 'Failed to fetch order summary' });
-    }
-  });
-  
-  /**
+/**
  * @swagger
  * /top-restaurants:
  *   get:
@@ -248,14 +127,7 @@ router.get('/order-summary', async (req, res) => {
  *       500:
  *         description: Failed to fetch top restaurants
  */
-router.get('/top-restaurants', async (req, res) => {
-  try {
-    const result = await executeQuery(topRestaurant);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch top restaurants' });
-  }
-});
+router.get("/top-restaurants", getTopRestaurt);
 
 /**
  * @swagger
@@ -269,15 +141,7 @@ router.get('/top-restaurants', async (req, res) => {
  *       500:
  *         description: Failed to fetch busiest delivery boy
  */
-
-router.get('/busiest-delivery-boy', async (req, res) => {
-  try {
-    const result = await executeQuery(getBusiestDelivery_Boy);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch busiest delivery boy' });
-  }
-});
+router.get("/busiest-delivery-boy", getBusiestDeliveryBoy);
 
 /**
  * @swagger
@@ -291,16 +155,7 @@ router.get('/busiest-delivery-boy', async (req, res) => {
  *       500:
  *         description: Failed to fetch frequent orders
  */
-
-
-router.get('/frequent-orders', async (req, res) => {
-  try {
-    const result = await executeQuery(getfrequentOrderDetails);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch frequent orders' });
-  }
-});
+router.get("/frequent-orders", getFrequestOrder);
 
 /**
  * @swagger
@@ -322,17 +177,7 @@ router.get('/frequent-orders', async (req, res) => {
  *       500:
  *         description: Failed to fetch order price
  */
-
-
-router.get('/specific-order-price/:orderId', async (req, res) => {
-  const { orderId } = req.params;
-  try {
-    const result = await executeQuery(getSpecificOrderPrice, [orderId]);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch specific order price' });
-  }
-});
+router.get("/specific-order-price/:orderId", getSpecificOrderById);
 
 /**
  * @swagger
@@ -354,19 +199,9 @@ router.get('/specific-order-price/:orderId', async (req, res) => {
  *       500:
  *         description: Failed to fetch delivery boy details
  */
+router.get("/order/:orderId/delivery-boy", getDeliveryBoyByOrderId);
 
-
-router.get('/order/:orderId/delivery-boy', async (req, res) => {
-    const { orderId } = req.params;
-    try {
-      const result = await executeQuery(getDeliveryBoyAssignedToSpecificOrder, [orderId]);
-      res.json(result);
-    } catch (error) {
-        res.status(500).send({ error: 'Failed to fetch delivery boy details for the order' });
-    }
-  });
-
-  /**
+/**
  * @swagger
  * /specific-order-details/{orderId}:
  *   get:
@@ -386,18 +221,7 @@ router.get('/order/:orderId/delivery-boy', async (req, res) => {
  *       500:
  *         description: Failed to fetch order details
  */
-
-
-
-router.get('/specific-order-details/:orderId', async (req, res) => {
-  const { orderId } = req.params;
-  try {
-    const result = await executeQuery(getSpecificOrdersDetails, [orderId]);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch specific order details' });
-  }
-});
+router.get("/specific-order-details/:orderId", getOrderDetailsByOrderId);
 
 /**
  * @swagger
@@ -411,16 +235,7 @@ router.get('/specific-order-details/:orderId', async (req, res) => {
  *       500:
  *         description: Failed to fetch out-for-delivery orders
  */
-
-
-router.get('/out-for-delivery-orders', async (req, res) => {
-  try {
-    const result = await executeQuery(getOutOfDeliveryOrder);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch out-for-delivery orders' });
-  }
-});
+router.get("/out-for-delivery-orders", getOutForDeliveryDetails);
 
 /**
  * @swagger
@@ -442,17 +257,7 @@ router.get('/out-for-delivery-orders', async (req, res) => {
  *       500:
  *         description: Failed to fetch delivery boy orders
  */
-
-
-router.get('/delivery-boy-orders/:id', async (req, res) => {
-  const deliveryBoyId = req.params.id;
-  try {
-    const result = await executeQuery(getOrderAssignedDeliverBoy, [deliveryBoyId]);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch delivery boy orders' });
-  }
-});
+router.get("/delivery-boy-orders/:id", assignedDeliveryBoyByorderId);
 
 /**
  * @swagger
@@ -474,17 +279,7 @@ router.get('/delivery-boy-orders/:id', async (req, res) => {
  *       500:
  *         description: Failed to fetch total sales for restaurant
  */
-
-
-router.get('/total-sales-for-restaurant/:restaurantId', async (req, res) => {
-  const { restaurantId } = req.params;
-  try {
-    const result = await executeQuery(getTotalSalesForRestaurant, [restaurantId]);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch total sales for restaurant' });
-  }
-});
+router.get("/total-sales-for-restaurant/:restaurantId",getTotalSalesByRestaurantId);
 
 /**
  * @swagger
@@ -498,14 +293,7 @@ router.get('/total-sales-for-restaurant/:restaurantId', async (req, res) => {
  *       500:
  *         description: Failed to fetch delivery boy status
  */
-router.get('/delivery-boy', async (req, res) => {
-  try {
-    const result = await executeQuery(getDeliveryBoyStatus);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch delivery boy status' });
-  }
-});
+router.get("/delivery-boy", getDeliveryBoy);
 
 /**
  * @swagger
@@ -519,15 +307,7 @@ router.get('/delivery-boy', async (req, res) => {
  *       500:
  *         description: Failed to fetch active delivery boys
  */
-
-router.get('/active-delivery-boys', async (req, res) => {
-  try {
-    const result = await executeQuery(getActiveDeliveryBoys);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch active delivery boys' });
-  }
-});
+router.get("/active-delivery-boys", getActiveDeliveryBoy);
 
 /**
  * @swagger
@@ -549,15 +329,6 @@ router.get('/active-delivery-boys', async (req, res) => {
  *       500:
  *         description: Failed to fetch delivery boy details for specific order
  */
-router.get('/delivery-boy-details-for-order/:orderId', async (req, res) => {
-  const { orderId } = req.params;
-  try {
-    const result = await executeQuery(getDeliveryBoyDetailsForSpecificOrder, [orderId]);
-    res.json(result);
-  } catch (error) {
-    res.status(500).send({ error: 'Failed to fetch delivery boy details for specific order' });
-  }
-});
-
+router.get("/delivery-boy-details-for-order/:orderId",getdeliveryboyDetails);
 // Exporting the router
 module.exports = router;
